itechart_project/requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
iniconfig==1.1.1
kombu==5.2.3
packaging==21.3
pbr==5.8.0
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytz==2021.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
uWSGI==2.0.20
vine==5.0.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
wrapt==1.13.3
